Changes made while repository was unavailable.

+++ include/base/types.h
@@ -82,15 +82,15 @@ namespace Ariadne {
   /*! \brief The type of a machine byte. */
   typedef unsigned char byte_type; 
   /*! \brief An unsigned integral type used to represent the size of a list. */
-  typedef size_t size_type;
+  typedef uint size_type;
   /*! \brief An signed integral type used to represent the position in a list with positive and negative indices. */
-  typedef ptrdiff_t index_type;
+  typedef int index_type;
   /*! \brief The type used for a unique identifier or key. */
-  typedef unsigned int id_type;
+  typedef uint id_type;
   /*! \brief An unsigned integral type used to represent a coordinate in state space. */
-  typedef unsigned short dimension_type;
+  typedef uint dimension_type;
   /*! \brief An integral type used to represent the smoothness of a map. */
-  typedef short smoothness_type;
+  typedef int smoothness_type;
   /*! \brief The type used to describe evolution time for discrete time systems. */
   typedef Numeric::Integer discrete_time_type;
   /*! \brief The type used to describe evolution time. */
